+++ angular_forms/lib/src/directives/shared.dart
@@ -57,8 +57,8 @@ void _throwError(AbstractControlDirective dir, String message) {
 
 ValidatorFn composeValidators(List<dynamic> validators) {
   return validators != null
-      ? Validators
-          .compose(validators.map<ValidatorFn>(normalizeValidator).toList())
+      ? Validators.compose(
+          validators.map<ValidatorFn>(normalizeValidator).toList())
       : null;
 }