+++ angular_test/CHANGELOG.md
@@ -1,10 +1,10 @@
 ## 2.0.0-beta+2
 
-*   Maintenence release: Stop using `SCREAMING_CAPS` consts from `dart:*`.
+*   Maintanence release: Stop using `SCREAMING_CAPS` consts from `dart:*`.
 
 ## 2.0.0-beta+1
 
-*   Maintenence release; declare official support for the Dart2 SDK.
+*   Maintanence release; declare official support for the Dart2 SDK.
 
 ## 2.0.0-beta